+++ src/cli/dcos_docker/commands/inspect_cluster.py
@@ -3,7 +3,8 @@ Tools for inspecting existing clusters.
 """
 
 import json
-from typing import Any, Dict  # noqa: F401
+from typing import Any  # noqa: F401
+from typing import Dict  # noqa: F401
 
 import click