+++ src/cli/dcos_docker/commands/mac_network.py
@@ -9,17 +9,8 @@ from io import BytesIO
 from pathlib import Path
 from shutil import copy, copytree, rmtree
 from tempfile import TemporaryDirectory
-from typing import (  # noqa: F401
-    Any,
-    Callable,
-    Dict,
-    Iterable,
-    List,
-    Optional,
-    Set,
-    Tuple,
-    Union,
-)
+from typing import Any  # noqa: F401
+from typing import Union
 
 import click
 import click_spinner
@@ -28,7 +19,7 @@ import docker
 from ._common import docker_client
 
 # We start these names with "e2e" rather than "dcos-e2e" to avoid a conflict
-# with "make clean".
+# with "dcos-docker clean".
 _PROXY_CONTAINER_NAME = 'e2e-proxy'
 _OPENVPN_CONTAINER_NAME = 'e2e-openvpn'