+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Next
 ----
 
 - Add ``dcos-docker create-loopback-sidecar`` and ``dcos-docker destroy-loopback-sidecar`` commands to provide unformatted block devices to DC/OS.
+- Add ``dcos-docker clean`` command to clean left over artifacts.
 
 2018.10.17.1
 ------------